packages in test folder refactored. Also added new test case
